// ncs_project-angular/src/app/model/question.ts
export class Question {
    questionNumber!: number;
    questionString?: string;
    questionCategory?: string;
    questionMarks?: number;
    questionOptionOne?: string;
    questionOptionTwo?: string;
    questionOptionThree?: string;
    questionOptionFour?: string;
    correctAnswer?: string;
}